requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.5.2
packaging==24.1
psycopg==3.1.19
psycopg2==2.9.9
pydot==2.0.0
PyJWT==2.8.0
pyparsing==3.1.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.12.0
uritemplate==4.1.1
